187: Update architecture source for stivale2
