+++ sysroot/usr/include/kernel/arch/arch.hpp
@@ -13,23 +13,39 @@
 #include <stdint.h>
 #include <stddef.h>
 
-const char* px_cpu_get_vendor();
-const char* px_cpu_get_model();
-
 struct registers;
 typedef struct registers registers_t;
-typedef void (*isr_t)(registers_t *);
 
+/* Shared i386 & amd64 */
 #if defined(__i386__) | defined(__i686__) | defined(__amd64__) | defined(__x86_64__)
-// Shared library code for the x86 family
+// Externed architecture variables
+extern const char* px_exception_descriptions[];
+// Architecture typedefs
+typedef void (*isr_t)(registers_t *);
+typedef struct stivale2_struct stivale2_info;
+// Architecture inline functions
+static inline void px_arch_cpuid(int flag, unsigned long eax, unsigned long ebx, unsigned long ecx, unsigned long edx)
+{
+    __asm__ volatile ("cpuid" : "=a"(eax), "=b"(ebx), "=c"(ecx), "=d"(edx) : "a"(flag));
+}
+static inline int px_arch_cpuid(int flag, int regs[4]) {
+    // ECX and EDX are swapped in order to make the strings readable
+    __asm__ volatile ("cpuid" : "=a"(*regs), "=b"(*(regs+1)), "=c"(*(regs+2)), "=d"(*(regs+3)) : "a"(flag));
+    return (int)regs[0];
+}
+// Kernel entry point
+extern "C" void px_kernel_main(stivale2_info *info);
+// i386+ & amd64 functions
+const char* px_cpu_get_vendor();
+const char* px_cpu_get_model();
+// Shared library code for i386+ & amd64 family
 #include <cpuid.h>
+#include <stivale2.h>
 #include <arch/i386/gdt.hpp>
 #include <arch/i386/idt.hpp>
 #include <arch/i386/isr.hpp>
 #include <arch/i386/timer.hpp>
 #include <arch/i386/ports.hpp>
-
-extern const char* px_exception_descriptions[];
 
 /**
  * @brief x86 BIOS based VGA pointers and data.
